Remove localStorage helper module
